requirements.txt
django
psycopg2-binary
psycopg2
djangorestframework
pillow
requests
postgres
